// src/common/enum/configkeys.enum.ts
/**
 * Enum para las opciones de configuración provenientes del .env
 *
 * En teoría se tiene que sincronizar el contenido de este archivo con la
 * validación en app.module.ts y obviamente el archivo .env|dotenv
 */
export enum ConfigKeys {
  JWT_REMEMBERME_EXPIRATION = 'JWT_REMEMBERME_EXPIRATION',
  STRIPE_SECRET = 'STRIPE_SECRET',
  PX_LAB_SERVER = 'PX_LAB_SERVER',
  REDIS_HOST = 'REDIS_HOST',
  REDIS_PASSWORD = 'REDIS_PASSWORD',
  REDIS_PORT = 'REDIS_PORT',
  REDIS_DB = 'REDIS_DB',
  REDIS_PREFIX = 'REDIS_PREFIX',
  ENABLE_SYSLOG = 'ENABLE_SYSLOG',
  MORGAN_TYPE = 'MORGAN_TYPE',
  JWT_SECRET = 'JWT_SECRET',
  JWT_EXPIRATION = 'JWT_EXPIRATION',
  API_ROUTE = 'API_ROUTE',
  CREATE_USERS = 'CREATE_USERS',
  FIRST_PASSWORD = 'FIRST_PASSWORD',
  MYSQL_DB = 'MYSQL_DB',
  MYSQL_HOST = 'MYSQL_HOST',
  MYSQL_USER = 'MYSQL_USER',
  MYSQL_PORT = 'MYSQL_PORT',
  MYSQL_PASSWORD = 'MYSQL_PASSWORD',
  API_SWAGGER = 'API_SWAGGER',
  PORT = 'PORT',
  SEND_USER_EMAILS = 'SEND_USER_EMAILS',
  SMTP_IGNORE_TLS = 'SMTP_IGNORE_TLS',
  SMTP_HOST = 'SMTP_HOST',
  SMTP_SECURE = 'SMTP_SECURE',
  SMTP_PORT = 'SMTP_PORT',
  SMTP_USER = 'SMTP_USER',
  SMTP_PASSWORD = 'SMTP_PASSWORD',
  SMTP_FROM_NAME = 'SMTP_FROM_NAME',
  SMTP_FROM_EMAIL = 'SMTP_FROM_EMAIL',
  EMAILS_INBOX = 'EMAILS_INBOX',
  SITE_NAME = 'SITE_NAME',
  BASE_URL = 'BASE_URL',
  SENTRY_DSN = 'SENTRY_DSN',
  NODE_ENV = 'NODE_ENV',
  TWILIO_ID = 'TWILIO_ID',
  TWILIO_KEY = 'TWILIO_KEY',
  TWILIO_NUMBER = 'TWILIO_NUMBER',
  FILE_ROUTE = 'FILE_ROUTE',
  AVATAR_SIZE = 'AVATAR_SIZE',
  API_KEY = 'API_KEY',

  GELF_SERVER_IP = 'GELF_SERVER_IP',
  GELF_PORT = 'GELF_PORT',
  GELF_FACILITY = 'GELF_FACILITY',

  WACHABOT_TOKEN = 'WACHABOT_TOKEN',
  WACHABOT_URL = 'WACHABOT_URL',
  WACHABOT_PREFIX = 'WACHABOT_PREFIX',
  WACHABOT_EMPRESA = 'WACHABOT_EMPRESA',
}